Refactor cmd-mox helpers into shared conftest
